+++ src/pages/QuoteBuilderCombined/QuoteDetail/QuoteProcess/index.tsx
@@ -204,6 +204,7 @@ export default function QuoteProcess(props) {
         message: null,
     });
     const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
+    const [fullScreen, setFullScreen] = useState(isMobile || isTablet);
 
     useEffect(() => {
         if (currentVersion !== 0) {
@@ -1631,7 +1632,7 @@ export default function QuoteProcess(props) {
                         setShowAiDialog(false);
                     }}
                     fullWidth
-                    fullScreen={isMobile || isTablet}
+                    fullScreen={fullScreen || (isMobile || isTablet)}
                     TransitionComponent={CustomDialogTransition}
                 >
                     <CustomDialogHeader
@@ -1639,6 +1640,11 @@ export default function QuoteProcess(props) {
                         onClose={() => {
                             setShowAiDialog(false);
                         }}
+                        isMinimized={!fullScreen}
+                        onMinimizeMaximize={() => {
+                            setFullScreen(prevState => !prevState)
+                        }}
+                        showManimizeMaximize={true}
                     />
                     <CustomDialogContent>
                         <div className="text-align-center">
@@ -1671,16 +1677,22 @@ export default function QuoteProcess(props) {
             {sendEmail && (
                 <Dialog
                     open={sendEmail}
-                    fullScreen={isMobile || isTablet}
+                    fullScreen={fullScreen || (isMobile || isTablet)}
                     TransitionComponent={CustomDialogTransition}
                     aria-labelledby="customized-dialog-title"
                     maxWidth="md"
-                    onClose={() => setSendEmail(false)}
+                    onClose={() => {
+                        setSendEmail(false)
+                        setFullScreen(false);
+                    }}
                     fullWidth
                 >
                     <CreateEmail
                         generatingFile={generatingPdfFile}
-                        handleClose={() => setSendEmail(false)}
+                        handleClose={() => {
+                            setSendEmail(false)
+                            setFullScreen(false);
+                        }}
                         fetchData={onSendEmailSuccess}
                         id={quoteData._id}
                         showESign={true}
@@ -1693,6 +1705,11 @@ export default function QuoteProcess(props) {
                         subject={`${user?.user?.brandName ?? "Brand"} Offer - ${quoteData?.quoteName ?? ""
                             }`}
                         fromQuote={true}
+                        isMinimized={!fullScreen}
+                        onMinimizeMaximize={() => {
+                            setFullScreen(prevState => !prevState)
+                        }}
+                        showManimizeMaximize={true}
                     />
                 </Dialog>
             )}