+++ src/pages/ReceivingTicket/ManageReceivingTicket/index.tsx
@@ -35,6 +35,7 @@ const ManageReceivingTicket = ({ isClone, receivingTicketId, productInventoryFor
   const [showConfirmDialog, setShowConfirmDialog] = useState(false);
   const [formsData, setFormsData] = useState([]);
   const [formValues, setFormValues] = useState({});
+  const [fullScreen, setFullScreen] = useState(isMobile || isTablet);
 
   useEffect(() => {
     setFormsData(setFieldsInAscendingOrder(receivingTicketData.fields));
@@ -140,7 +141,7 @@ const ManageReceivingTicket = ({ isClone, receivingTicketId, productInventoryFor
       axiosInstance()
         .post(`${receivingTicket.receivingTicketApi}`, values)
         .then(({ data: { data, message } }) => {
-          if(isRedirectToDetailPage){
+          if (isRedirectToDetailPage) {
             history.push(`${routes.receivingTicketDetail.path}/${data._id}`);
           }
           setSubmitting(false);
@@ -182,7 +183,7 @@ const ManageReceivingTicket = ({ isClone, receivingTicketId, productInventoryFor
       <Dialog
         maxWidth="md"
         fullWidth
-        fullScreen={isMobile || isTablet}
+        fullScreen={fullScreen || (isMobile || isTablet)}
         TransitionComponent={CustomDialogTransition}
         aria-labelledby="customized-dialog-title"
         onClose={(e, reason) => {
@@ -198,6 +199,11 @@ const ManageReceivingTicket = ({ isClone, receivingTicketId, productInventoryFor
             if (isFieldNotTouched(receivingTicketData, formValues)) onClose();
             else setShowConfirmDialog(true);
           }}
+          isMinimized={!fullScreen}
+          onMinimizeMaximize={() => {
+            setFullScreen(prevState => !prevState)
+          }}
+          showManimizeMaximize={true}
         />
         {loading || !receivingTicketData.fields.length ? (
           <>